+++ ext/MeshIntegralsEnzymeExt.jl
@@ -13,6 +13,7 @@ function MeshIntegrals.jacobian(
     if Dim != length(ts)
         throw(ArgumentError("ts must have same number of dimensions as geometry."))
     end
+    
     return Meshes.to.(Enzyme.jacobian(Enzyme.Forward, geometry, ts...))
 end